Removed buffer block/unblock and FINALLY removed the reader/writer conditions from buffer! WOOHOO!  This saves nearly 50% of the original buffer size, reducing it down to 88 bytes!  I'm very pleased with this update.  Need to update the readme and stuff, and tag some fixes.

+++ src/list.c
@@ -739,9 +739,10 @@ int list__update(List *list, Buffer **callers_buf, void *data, uint32_t size, ui
   // Now that we're done reading buf, we need to change the caller's buf via indirection so they see the new buffer.
   *callers_buf = new_buffer;
 
-  // Check to see if this was just a raw-to-raw update.
+  // Check to see if this was just a raw-to-raw update.  Compressors update list size on their own.
   if((buf->flags & compressing) == 0)
-    __sync_fetch_and_add(&list->current_raw_size, (size - buf->data_length));
+    // Coerce to allow a negative value to the atomic; otherwise an underflow can be sent.
+    __sync_fetch_and_add(&list->current_raw_size, (int)(size - buf->data_length));
 
   // Mark the buffer dirty, remove the updating flag, and throw it in the dirty pool for future eviction.
   pthread_mutex_lock(&buf->lock);
